pyproject.toml
[tool.poetry]
name = "connectathon-sep-2021"
version = "0.1.0"
description = ""
authors = ["Geoff Low <glow@mdsol.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
"fhir.resources" = "^6.1.0"
jupyterlab = "^3.1.12"
requests = "^2.26.0"
flat-table = "^1.1.1"
fhirclient = "^3.2.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
